Fix the HTML to validate with W3C Markup Validation Service (#7)
* Remove the redundant http-equiv charset meta element from head.html

The element has been replaced with <meta charset="UTF-8" /> in HTML5, which
is also present in the head.html file, right next to the former. There is no
need to have both listed together in the document HEAD.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Remove the invalid http-equiv date meta element from head.html

Date is not a valid value to be used with the http-equiv attribute. Delete the
whole invalid element and move the date format to the meta name element below
it.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Add missing alt attributes to images on the Home page

alt attributes are required both for accessibility and also for HTML code
validation by W3C Markup Validation Service.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Remove the invalid p element from the Home page

Fix the Hugo configuration, so that the superfluous p elements are not created
around the sponsor logo. The Home page now validates correctly with the W3C
Markup Validation Service.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Remove the invalid p and li elements from the Foundation page

Modify the configuration not to create superfluous and invalid paragraph and
list item elements.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Fix the invalid netlify form attribute on the Foundations page

Use data-netlify instead of just netlify, which is an invalid form attribute.
The Foundations page now validates with the W3C Markup Validation Service.

See: https://docs.netlify.com/forms/setup/#html-forms

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Add missing alt attribute to the image on the Donations page

alt attributes are required both for accessibility and also to validate the
HTML code with the W3C Markup Validation Service.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Remove the invalid i element from the Donations page

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Remove the redundant type attributes from script elements on the Donations page

The script element does not require these kind of type attributes in HTML5.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Fix the invalid label for attribute on the Donations page

The invalid for attribute makes the label non-functional.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>

* Update the Netlify build configuration to Hugo 0.62.2

Use the newer version of Hugo in order to fix invalid HTML created by Netlify,
i.e. the duplicated a mailto elements on the Foundation and Donation pages.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>