don't recommend adding system-wide repository key
Current Debian guidelines discourage adding system-wide signing
keys for repositories like ours. The reasoning is that in case
such a key is compromised it can be used to sign other
repositories and completely hijack the system, while the new
way of doing things guarantees that the key will only be
considered for our repository, so the attacker must compromise
both the key and the repo to achieve anything, which is more
hassle for the attacker, hence the system is more secure.